added services and dbContext

+++ BooksWebApi/DTO/GetBookDto.cs
@@ -6,5 +6,5 @@ public class GetBookDto
     public string Title { get; set; } = string.Empty;
     public string Author { get; set; } = string.Empty;
     public decimal Rating { get; set; }
-    public decimal ReviewsNumber { get; set; } 
+    public int ReviewsNumber { get; set; } 
 }
